Correction de la route deletePost

+++ backend/controllers/post.js
@@ -10,11 +10,11 @@ const userID = (req) => {
     return id;
 }
 
-// Créer un post
+// Créer un post -----
 exports.createPost = (req, res, next) => {
     let imgUrl;
     if (!req.file) {
-        imgUrl = null;
+        return;
     } else {
         imgUrl = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`;
     }
@@ -27,7 +27,7 @@ exports.createPost = (req, res, next) => {
     .catch(error => res.status(400).json({error}));
 }
 
-// Modifier un post
+// Modifier un post -----
 exports.editPost = (req, res, next) => {
     db.Post.findOne({
         where: {id: req.params.id},
@@ -73,25 +73,49 @@ exports.editPost = (req, res, next) => {
     .catch(error => res.status(500).json({error}))
 }
 
-// Supprimer un post
+// Supprimer un post -----
 exports.deletePost = (req, res, next) => {
+    // On recherche le post dont l'id est dans les paramètres de la requête
     db.Post.findOne({
-        where: {id: req.params.id}
+        where: {id: req.params.id},
+        include: [
+            {
+                model: db.User,
+                attributes: ["id", "firstname", "lastname", "email", "department", "profilePicture", "isAdmin"]
+            },
+        ]
     })
     .then(post => {
-        const filename = post.media.split('/images/')[1];
-        fs.unlink(`images/${filename}`, () => {
-            db.Post.destroy({
-                where: {id: req.params.id}
+        // Si le post existe
+        if (post) {
+            db.User.findOne({
+                where: {id: userID(req)}
             })
-            .then(() => res.status(200).json({message: "Post supprimé !"}))
-            .catch(error => res.status(400).json({error}));
-        })
+            .then(admin => {
+                if (post.userId == userID(req) || admin.isAdmin === true) {
+                    const filename = post.media.split('/images/')[1];
+                    fs.unlink(`images/${filename}`, () => {
+                        db.Post.destroy({
+                            where: {id: req.params.id}
+                        })
+                        .then(() => res.status(200).json({message: "Post supprimé !"}))
+                        .catch(error => res.status(400).json({error}));
+                    })
+                } else {
+                    res.status(403).json({message: "Vous n'êtes pas autorisé(e) à supprimer ce post !"})
+                }
+            })
+            .catch(error => res.status(500).json({error}))
+        }
+        // Si le post n'existe pas 
+        else {
+            res.status(404).json({erreur: "Le post n'existe pas!"})
+        }
     })
     .catch(error => res.status(500).json({error}));
 }
 
-// Afficher tous les posts
+// Afficher tous les posts -----
 exports.getAllPosts = (req, res, next) => {
     db.Post.findAll({
         attributes: ['id', 'title', 'media', 'likes', 'dislikes', 'createdAt'],
@@ -113,7 +137,7 @@ exports.getAllPosts = (req, res, next) => {
     .catch(error => res.status(400).json({error}));
 }
 
-// Afficher un post
+// Afficher un post -----
 exports.getOnePost = (req, res, next) => {
     db.Post.findOne({
         attributes: ['id', 'title', 'media', 'likes', 'dislikes', 'createdAt'],
